Fix Python timings on ARCHER2
Contrary to what was written in the (manually prepared) `python-weak.csv` file,
the number of iterations used was 200, not 1000.  I reran the benchmarks with
the correct number of iterations, to get timings more consistently with the
other implementations.  Note that this is now about 4x slower than all the other
implementations, which also explains the good scaling efficiency: the baseline
is much higher.